+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 * 【core   】     FileTypeUtil使用长匹配优先（pr#1457@Github）
 * 【core   】     IterUtil和CollUtil增加isEqualList方法（issue#I3A3PY@Gitee）
 * 【crypto 】     增加PBKDF2（issue#1416@Github）
+* 【core   】     增加FuncKeyMap（issue#1402@Github）
 
 ### Bug修复
 * 【socket 】     修复Client创建失败资源未释放问题。